anpassung Demands und Owes

+++ ViewModels/BuildingViewModel.cs
@@ -21,7 +21,7 @@ namespace CleanHub.ViewModels
         }
         [Ignore]
         [IgnoreDataMember]
-        public int? ReserveTotal { get; set; }
+        public double? ReserveTotal { get; set; }
 
         public List<BuildingProductViewModel> BuildingProducts { get; set; } = null!;
         public ICollection<CustomerViewModel>? Customers { get; set; }
